Implement Bugs card for Test Case page. Refs #18
Note: this is missing the add/remove functionality for now.
Bug.remove is very powerful and can remove bugs that are
attached to a TC but not coming from executions and also
bugs that are coming from executions.

Adding bugs is the reverse operation and I'm not quite sure how
that functionality is used outside test runs!